teacher student subject 100%

+++ src/main/java/com/schoolmanagement/model/Mark.java
@@ -40,7 +40,7 @@ public class Mark {
 	private Integer semester;
 
 	@Column(name = "coefficient")
-	private Integer coefficient;
+	private Float coefficient;
 
 	@Column(name = "created_date")
 	private LocalDate createdDate;
