+++ src/main/java/com/schoolmanagement/model/request/MarkRequest.java
@@ -21,7 +21,7 @@ public class MarkRequest {
 
 	private Integer semester;
 
-	private List<Integer> coefficient;
+	private List<Float> coefficient;
 
 	private Integer createdBy;
